fix(node): restrict restart replay to processed blocks; move restore into internal/node
Headers are also persisted for pending blocks that were never verified
(store pending path), and they survive a later signature or state-
transition rejection. Replaying them after a restart lets an unverified
block connected to the justified subtree win the head tie-break and
break duties with a missing head state. Only blocks with a stored
post-state enter the restored array - the same processed-blocks
invariant the spec store maintains.

Startup fork-choice restore is consensus-critical Engine wiring, so it
moves from cmd/gean into internal/node, which also puts its regression
tests inside make test and the CI race pass; they previously ran in
neither.

Refs #351
