+++ cmd/gean/main_test.go
@@ -106,101 +106,6 @@ func TestInitStoreFromStatePersistsAnchor(t *testing.T) {
 	if s.GetState(root) == nil {
 		t.Fatal("expected persisted state")
 	}
-
-	fc, err := forkChoiceFromStore(s)
-	if err != nil {
-		t.Fatalf("fork choice from store: %v", err)
-	}
-	if fc.Len() != 1 || fc.NodeIndex(root) != 0 {
-		t.Fatalf("fresh store should anchor fork choice at head only: len=%d idx=%d", fc.Len(), fc.NodeIndex(root))
-	}
-}
-
-func TestForkChoiceFromStoreRejectsMissingHeader(t *testing.T) {
-	s := newTestStore()
-	s.SetHead([32]byte{0x01})
-
-	_, err := forkChoiceFromStore(s)
-	if err == nil {
-		t.Fatal("expected missing head header error")
-	}
-}
-
-func TestForkChoiceFromStoreReplaysChainToJustified(t *testing.T) {
-	s := newTestStore()
-	genesis := [32]byte{0xaa}
-	blockA := [32]byte{0xbb}
-	blockB := [32]byte{0xcc}
-	blockC := [32]byte{0xdd}
-
-	s.InsertBlockHeader(genesis, &types.BlockHeader{Slot: 0})
-	s.InsertBlockHeader(blockA, &types.BlockHeader{Slot: 2, ParentRoot: genesis})
-	s.InsertBlockHeader(blockB, &types.BlockHeader{Slot: 4, ParentRoot: blockA})
-	s.InsertBlockHeader(blockC, &types.BlockHeader{Slot: 6, ParentRoot: blockB})
-	s.SetHead(blockC)
-	s.SetLatestJustified(&types.Checkpoint{Root: blockA, Slot: 2})
-
-	fc, err := forkChoiceFromStore(s)
-	if err != nil {
-		t.Fatalf("fork choice from store: %v", err)
-	}
-	if fc.NodeIndex(blockA) < 0 {
-		t.Fatal("justified root missing from restored fork choice")
-	}
-	if head := fc.UpdateHead(blockA); head != blockC {
-		t.Fatalf("head=%x, want stored leaf %x", head, blockC)
-	}
-}
-
-func TestForkChoiceFromStoreFallsBackWhenJustifiedUnreachable(t *testing.T) {
-	s := newTestStore()
-	head := [32]byte{0x02}
-	s.InsertBlockHeader(head, &types.BlockHeader{Slot: 6, ParentRoot: [32]byte{0x03}})
-	s.SetHead(head)
-	s.SetLatestJustified(&types.Checkpoint{Root: [32]byte{0x09}, Slot: 2})
-
-	fc, err := forkChoiceFromStore(s)
-	if err != nil {
-		t.Fatalf("fork choice from store: %v", err)
-	}
-	if fc.Len() != 1 || fc.NodeIndex(head) != 0 {
-		t.Fatalf("expected fallback anchor at head: len=%d idx=%d", fc.Len(), fc.NodeIndex(head))
-	}
-}
-
-func TestForkChoiceFromStoreReplaysSiblingForks(t *testing.T) {
-	s := newTestStore()
-	justified := [32]byte{0xaa}
-	staleB := [32]byte{0xbb}
-	staleHead := [32]byte{0xcc}
-	forkB := [32]byte{0xdd}
-	forkTip := [32]byte{0xee}
-
-	s.InsertBlockHeader(justified, &types.BlockHeader{Slot: 2})
-	s.InsertBlockHeader(staleB, &types.BlockHeader{Slot: 4, ParentRoot: justified})
-	s.InsertBlockHeader(staleHead, &types.BlockHeader{Slot: 6, ParentRoot: staleB})
-	s.InsertBlockHeader(forkB, &types.BlockHeader{Slot: 3, ParentRoot: justified})
-	s.InsertBlockHeader(forkTip, &types.BlockHeader{Slot: 5, ParentRoot: forkB})
-	s.SetHead(staleHead)
-	s.SetLatestJustified(&types.Checkpoint{Root: justified, Slot: 2})
-
-	fc, err := forkChoiceFromStore(s)
-	if err != nil {
-		t.Fatalf("fork choice from store: %v", err)
-	}
-	if fc.NodeIndex(forkTip) < 0 {
-		t.Fatal("sibling fork missing from restored fork choice")
-	}
-
-	att := &types.AttestationData{Slot: 5, Head: &types.Checkpoint{Root: forkTip, Slot: 5}}
-	for vid := uint64(0); vid < 3; vid++ {
-		if !fc.SetKnownVote(vid, forkTip, 5, att) {
-			t.Fatalf("vote for fork tip rejected for validator %d", vid)
-		}
-	}
-	if head := fc.UpdateHead(justified); head != forkTip {
-		t.Fatalf("head=%x, want weighted fork tip %x", head, forkTip)
-	}
 }
 
 func TestBootstrapStoreRejectsPreDevnet5Database(t *testing.T) {
